// src/shared/types.ts
export interface IAuthSession {
  userId: number;
  tenantId: number;
  email: string;
  rankLevel: number;
  departmentId: number | null;
}